[wofi] Use a dark GTK theme when available
This commit configures wofi to use the dark version of the current GTK
theme if is available.

This change however has been made in sway config because wofi does not
have a proper config file as of right now.

This change will be moved to a proper wofi conf file later.